+++ client/src/components/ui/AdminSettingsDialog.tsx
@@ -1,19 +1,23 @@
-import { useEffect, useState } from 'react';
+import { useId, useEffect, useState } from 'react';
 import * as Ariakit from '@ariakit/react';
-import { ShieldEllipsis } from 'lucide-react';
 import { useForm, Controller } from 'react-hook-form';
+import { ChevronDown, ShieldEllipsis } from 'lucide-react';
 import { Permissions, SystemRoles } from 'librechat-data-provider';
 import {
-  OGDialog,
-  OGDialogTitle,
-  OGDialogContent,
-  OGDialogTrigger,
+  Label,
   Button,
   Switch,
+  OGDialog,
   DropdownPopup,
+  OGDialogTitle,
+  OGDialogHeader,
+  OGDialogFooter,
+  OGDialogContent,
+  OGDialogTrigger,
+  OGDialogDescription,
 } from '@librechat/client';
-import type { Control, UseFormSetValue, UseFormGetValues } from 'react-hook-form';
 import type { PermissionTypes } from 'librechat-data-provider';
+import type { Control } from 'react-hook-form';
 import type { TranslationKeys } from '~/hooks/useLocalize';
 import { useLocalize, useAuthContext, useRoleSelector } from '~/hooks';
 
@@ -37,7 +41,11 @@ export interface AdminSettingsDialogProps {
   mutation: {
     mutate: (data: { roleName: string; updates: Record<Permissions, boolean> }) => void;
     isLoading: boolean;
+    /** When it flips true the dialog closes itself */
+    isSuccess?: boolean;
   };
+  /** Localization key for the screen-reader description of the dialog */
+  descriptionKey?: TranslationKeys;
   /** Whether to show the admin access warning when ADMIN role and USE permission is displayed (default: true) */
   showAdminWarning?: boolean;
   /** Custom trigger element. If not provided, uses default button with icon and text */
@@ -57,29 +65,35 @@ export interface AdminSettingsDialogProps {
 }
 
 type LabelControllerProps = {
+  id: string;
   label: string;
   permission: Permissions;
   control: Control<FormValues, unknown, FormValues>;
-  setValue: UseFormSetValue<FormValues>;
-  getValues: UseFormGetValues<FormValues>;
   onConfirm?: (newValue: boolean, onChange: (value: boolean) => void) => void;
 };
 
 const LabelController: React.FC<LabelControllerProps> = ({
+  id,
   control,
   permission,
   label,
   onConfirm,
 }) => (
-  <div className="mb-4 flex items-center justify-between gap-2">
-    {label}
+  <div className="flex items-center justify-between gap-4 px-4 py-3.5">
+    <Label
+      htmlFor={id}
+      className="w-auto cursor-pointer select-none break-normal text-sm font-medium text-text-primary"
+    >
+      {label}
+    </Label>
     <Controller
       name={permission}
       control={control}
       render={({ field }) => (
         <Switch
           {...field}
-          checked={field.value}
+          id={id}
+          checked={field.value ?? false}
           onCheckedChange={(val) => {
             if (val === false && onConfirm) {
               onConfirm(val, field.onChange);
@@ -87,7 +101,7 @@ const LabelController: React.FC<LabelControllerProps> = ({
               field.onChange(val);
             }
           }}
-          value={field.value?.toString()}
+          value={(field.value ?? false).toString()}
           aria-label={label}
         />
       )}
@@ -101,6 +115,7 @@ const AdminSettingsDialog: React.FC<AdminSettingsDialogProps> = ({
   permissions,
   menuId,
   mutation,
+  descriptionKey,
   showAdminWarning = true,
   trigger,
   dialogContentClassName,
@@ -110,9 +125,14 @@ const AdminSettingsDialog: React.FC<AdminSettingsDialogProps> = ({
 }) => {
   const localize = useLocalize();
   const { user } = useAuthContext();
-  const { mutate, isLoading } = mutation;
+  const { mutate, isLoading, isSuccess } = mutation;
 
+  const idPrefix = useId();
+  const roleLabelId = useId();
+  const roleValueId = useId();
+  const [isDialogOpen, setIsDialogOpen] = useState(false);
   const [isRoleMenuOpen, setIsRoleMenuOpen] = useState(false);
+  const [wasSuccessful, setWasSuccessful] = useState(isSuccess);
   const {
     selectedRole,
     isSelectedCustomRole,
@@ -125,8 +145,6 @@ const AdminSettingsDialog: React.FC<AdminSettingsDialogProps> = ({
   const {
     reset,
     control,
-    setValue,
-    getValues,
     handleSubmit,
     formState: { isSubmitting },
   } = useForm<FormValues>({
@@ -141,9 +159,24 @@ const AdminSettingsDialog: React.FC<AdminSettingsDialogProps> = ({
     reset(defaultValues);
   }, [isSelectedCustomRole, isCustomRoleLoading, isCustomRoleError, defaultValues, reset]);
 
+  if (isSuccess !== wasSuccessful) {
+    setWasSuccessful(isSuccess);
+    if (isSuccess === true && isDialogOpen) {
+      setIsDialogOpen(false);
+    }
+  }
+
   if (user?.role !== SystemRoles.ADMIN) {
     return null;
   }
+
+  const handleDialogOpenChange = (open: boolean) => {
+    if (!open) {
+      setIsRoleMenuOpen(false);
+      reset(defaultValues);
+    }
+    setIsDialogOpen(open);
+  };
 
   const onSubmit = (data: FormValues) => {
     mutate({ roleName: selectedRole, updates: data });
@@ -163,82 +196,104 @@ const AdminSettingsDialog: React.FC<AdminSettingsDialogProps> = ({
 
   return (
     <>
-      <OGDialog>
+      <OGDialog open={isDialogOpen} onOpenChange={handleDialogOpenChange}>
         <OGDialogTrigger asChild>{trigger ?? defaultTrigger}</OGDialogTrigger>
         <OGDialogContent
-          className={
-            dialogContentClassName ??
-            'w-11/12 max-w-lg border-border-light bg-surface-dialog text-text-primary'
-          }
+          className={dialogContentClassName ?? 'w-11/12 max-w-2xl gap-0 overflow-hidden p-0'}
         >
-          <OGDialogTitle>
-            {localize('com_ui_admin_settings_section', { section: localize(sectionKey) })}
-          </OGDialogTitle>
-
-          {/* Role selection dropdown */}
-          <div className="flex items-center gap-2">
-            <span className="font-medium">{localize('com_ui_role_select')}:</span>
-            <DropdownPopup
-              unmountOnHide={true}
-              menuId={menuId}
-              isOpen={isRoleMenuOpen}
-              setIsOpen={setIsRoleMenuOpen}
-              trigger={
-                <Ariakit.MenuButton className="inline-flex min-w-[6rem] items-center justify-center rounded-lg border border-border-light bg-transparent px-2 py-1 text-text-primary transition-all ease-in-out hover:bg-surface-tertiary">
-                  {selectedRole}
-                </Ariakit.MenuButton>
-              }
-              items={roleDropdownItems}
-              itemClassName="items-center justify-center"
-              sameWidth={true}
-            />
-          </div>
-          {/* Permissions form */}
-          <form onSubmit={handleSubmit(onSubmit)}>
-            <div className="py-5">
-              {permissions.map(({ permission, labelKey }) => {
-                const label = localize(labelKey);
-                const needsConfirm =
-                  selectedRole === SystemRoles.ADMIN &&
-                  confirmPermissions.includes(permission) &&
-                  onPermissionConfirm;
-
-                return (
-                  <div key={permission}>
-                    <LabelController
-                      control={control}
-                      permission={permission}
-                      label={label}
-                      getValues={getValues}
-                      setValue={setValue}
-                      onConfirm={
-                        needsConfirm
-                          ? (newValue, onChange) =>
-                              onPermissionConfirm(permission, newValue, onChange)
-                          : undefined
-                      }
-                    />
-                    {showAdminWarning &&
-                      selectedRole === SystemRoles.ADMIN &&
-                      permission === Permissions.USE && (
-                        <div className="mb-2 max-w-full whitespace-normal break-words text-sm text-text-destructive">
-                          <span>{localize('com_ui_admin_access_warning')}</span>
-                          {'\n'}
-                          <a
-                            href="https://www.librechat.ai/docs/configuration/librechat_yaml/object_structure/interface"
-                            target="_blank"
-                            rel="noreferrer"
-                            className="text-link underline"
-                          >
-                            {localize('com_ui_more_info')}
-                          </a>
-                        </div>
-                      )}
-                  </div>
-                );
-              })}
+          <OGDialogHeader className="px-5 py-5 pr-14 text-left sm:px-6">
+            <div className="flex items-center gap-3">
+              <div className="flex size-10 shrink-0 items-center justify-center rounded-xl border border-border-light bg-surface-secondary text-text-secondary">
+                <ShieldEllipsis className="size-5" aria-hidden="true" />
+              </div>
+              <div className="min-w-0 space-y-1">
+                <OGDialogTitle className="text-xl leading-7">
+                  {localize('com_ui_admin_settings_section', { section: localize(sectionKey) })}
+                </OGDialogTitle>
+                {descriptionKey && (
+                  <OGDialogDescription className="sr-only">
+                    {localize(descriptionKey)}
+                  </OGDialogDescription>
+                )}
+              </div>
             </div>
-            <div className="flex justify-end">
+          </OGDialogHeader>
+
+          <form onSubmit={handleSubmit(onSubmit)}>
+            <div className="space-y-4 p-4 sm:p-6">
+              <div className="grid gap-3 rounded-xl border border-border-light bg-surface-secondary p-4 sm:grid-cols-[minmax(0,1fr)_9rem] sm:items-center">
+                <span id={roleLabelId} className="text-sm font-semibold text-text-primary">
+                  {localize('com_ui_role_select')}
+                </span>
+                <DropdownPopup
+                  unmountOnHide={true}
+                  menuId={menuId}
+                  isOpen={isRoleMenuOpen}
+                  setIsOpen={setIsRoleMenuOpen}
+                  trigger={
+                    <Ariakit.MenuButton
+                      aria-labelledby={`${roleLabelId} ${roleValueId}`}
+                      className="inline-flex h-10 w-full items-center justify-between gap-2 rounded-lg border border-border-medium bg-transparent px-3 text-sm text-text-primary transition-colors hover:bg-surface-hover focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-text-primary"
+                    >
+                      <span id={roleValueId} className="truncate font-medium">
+                        {selectedRole}
+                      </span>
+                      <ChevronDown
+                        className="size-4 shrink-0 text-text-secondary"
+                        aria-hidden="true"
+                      />
+                    </Ariakit.MenuButton>
+                  }
+                  items={roleDropdownItems}
+                  itemClassName="items-center justify-center"
+                  sameWidth={true}
+                />
+              </div>
+
+              <div className="divide-y divide-border-light overflow-hidden rounded-xl border border-border-light bg-surface-secondary">
+                {permissions.map(({ permission, labelKey }) => {
+                  const label = localize(labelKey);
+                  const needsConfirm =
+                    selectedRole === SystemRoles.ADMIN &&
+                    confirmPermissions.includes(permission) &&
+                    onPermissionConfirm;
+
+                  return (
+                    <div key={permission}>
+                      <LabelController
+                        id={`${idPrefix}-${permission}`}
+                        control={control}
+                        permission={permission}
+                        label={label}
+                        onConfirm={
+                          needsConfirm
+                            ? (newValue, onChange) =>
+                                onPermissionConfirm(permission, newValue, onChange)
+                            : undefined
+                        }
+                      />
+                      {showAdminWarning &&
+                        selectedRole === SystemRoles.ADMIN &&
+                        permission === Permissions.USE && (
+                          <div className="whitespace-normal break-words border-t border-border-light px-4 py-3 text-sm text-text-destructive">
+                            <span>{localize('com_ui_admin_access_warning')}</span>{' '}
+                            <a
+                              href="https://www.librechat.ai/docs/configuration/librechat_yaml/object_structure/interface"
+                              target="_blank"
+                              rel="noreferrer"
+                              className="text-link underline"
+                            >
+                              {localize('com_ui_more_info')}
+                            </a>
+                          </div>
+                        )}
+                    </div>
+                  );
+                })}
+              </div>
+            </div>
+
+            <OGDialogFooter className="bg-transparent px-4 py-4 sm:px-6">
               <Button
                 type="submit"
                 variant="submit"
@@ -247,11 +302,12 @@ const AdminSettingsDialog: React.FC<AdminSettingsDialogProps> = ({
                   isLoading ||
                   (isSelectedCustomRole && (isCustomRoleLoading || isCustomRoleError))
                 }
+                className="font-bold"
                 aria-label={localize('com_ui_save')}
               >
                 {localize('com_ui_save')}
               </Button>
-            </div>
+            </OGDialogFooter>
           </form>
         </OGDialogContent>
       </OGDialog>
